backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-CORS
psycopg2-binary
python-dotenv
requests
Werkzeug

